requirements.txt
flask
python-dotenv
google-generativeai
gspread
google-api-python-client
requests
